app/requirements.txt
sentencepiece==0.1.97
onnx==1.13.0
onnxruntime-gpu==1.13.1
optimum==1.6.1
fastapi[all]==0.89.1